Allow for a blank space in SSplotData for users to plot more later.
* Added a period to the caption of the plot.

* separate out code to determine plotting colours for the fleets,
this allows the same colours to be used outside of the function
without needing to recode the ifelse statements.

* Add an empty datatype of "empty" that will allow users to plot
additional data after the plot is generated. The empty space is
not included by default with the "all" case, only when called for
specifically in datatypes = c('xxx', 'empty')

* fix a bug where ntypes is based on what is left after the subset,
i tested this with the sablefish model in that not every fleet has
all types of data, which led to the itype having larger numbers than
what was left in the typetable after subsetting for fleet. Now, instead
of using the maximum value, which might have gaps, just the number
of unique values is reported.

* use the absolute value of the range of years so that the empty
type can have a negative year, which makes it not visible on the plot,
the negative years as entered in the dat file are positive in this table
b/c the "use" column takes over the negative business. I had to use
a negative value b/c NA values aren't allowed in the symbol.

* don't plot a fleet name if the type is empty